requirements.txt
asyncpg>=0.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
pyarrow>=12.0.0  # for parquet support 